Move getGenesisStateProof to block.js.
